pyproject.toml
[project]
name = "doaz-eng-rag"
version = "0.1.0"
description = "Doaz Engineering RAG System"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "qdrant-client>=1.12.0",
    "openai>=1.57.0",
    "anthropic>=0.40.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.18",
    "loguru>=0.7.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.0",
    "cohere>=5.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
